+++ src/client/ui/ui-toolbar.js
@@ -13,26 +13,19 @@ class UiToolbar extends UiWidget {
         this.opened = false
         this.entries = options.entries
 
-        if (READ_ONLY) {
+        this.menu.bindShortcuts(this.entries)
 
-            this.container.style.display = 'none'
+        this.container.addEventListener('fast-click', (e)=>{
+            if (this.opened) this.close()
+            else this.open(e)
+        })
 
-        } else {
-
-            this.menu.bindShortcuts(this.entries)
-
-            this.container.addEventListener('fast-click', (e)=>{
-                if (this.opened) this.close()
-                else this.open(e)
+        this.menu.on('close', ()=>{
+            setTimeout(()=>{
+                this.opened = false
+                this.toggleState()
             })
-
-            this.menu.on('close', ()=>{
-                setTimeout(()=>{
-                    this.opened = false
-                    this.toggleState()
-                })
-            })
-        }
+        })
 
     }