+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - bug fixes
   - server: `fullscreen` option not working when `read-only` is set
   - led and text widgets not client-syncing  
+  - patchbay: address not working properly when set to `auto`
 
 ##  1.27.0